+++ crates/ltx_compiler/src/lib.rs
@@ -1,14 +1,9 @@
-pub fn add(left: u64, right: u64) -> u64 {
-    left + right
-}
+//! Ltx compiler which compiles Ltx source code into PDF Via rust lib tectonic.
+//! Or via pdflatex or LuaTeX using UNIX processes.
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn it_works() {
-        let result = add(2, 2);
-        assert_eq!(result, 4);
-    }
-}
+/// will handles the compilation of Ltx source code into HTML or PDF.
+pub mod build;
+/// returns the configuration for the compiler [`name`, `engine_name`, `engine_args`].
+pub mod config;
+/// will watch for changes in the source code and recompile as needed.
+pub mod watch;